requirements.txt
python-lambda
python-googl
retrying
